removed stellargradebook.py and added get_gradebook_id() to base.py
